% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieval-fxns.R
\name{get_bootstrap_trees}
\alias{get_bootstrap_trees}
\title{Download zipped bootstrap trees from the OSF repository, unzip a desired number
of bootstrapped trees to a output directory}
\usage{
get_bootstrap_trees(
  all_trees = TRUE,
  num_trees = NA,
  output_dir,
  tree_type = "dated",
  setSeed = TRUE,
  randomSeed = 111,
  osf_token = NA
)
}
\arguments{
\item{all_trees}{Boolean. If true, gather all bootstraps available, if FALSE, use num_trees to designate desired number of trees}

\item{num_trees}{Numeric. Specify the number of bootstraps to gather, max is all trees (1004)}

\item{output_dir}{Character string. Specify the location for where bootstrap tree directry will be stored}

\item{tree_type}{Character string. Either "dated" or "undated" bootstraps}

\item{setSeed}{Boolean. If you are specifying a number of bootstraps, you can set this to TRUE and specify randomSeed to ensure reproducible bootstrap subsets are gathered}

\item{randomSeed}{numeric. Reproducible seed to set if setSeed is TRUE}

\item{osf_tokjen}{Character string. The string that authorizes use of the OSF repo, this will be removed once the data is public}
}
\description{
Download zipped bootstrap trees from the OSF repository, unzip a desired number
of bootstrapped trees to a output directory
}
